requirements.txt
flask>=3.0
PyMuPDF>=1.24
openpyxl>=3.1
anthropic>=0.40
python-dotenv>=1.0
requests
gunicorn
